# Performance improvement for bytes-type Column searchsorted