@mhvk 2021-12-31
@taldcroft - the slow-down with `__array_function__` is indeed generic...  But for this case, I think we do not need `__array_function__`: `np.searchsorted` should defer to the `.searchsorted` method if it is present, so you can just define that!