@taldcroft 2022-01-01
> np.searchsorted should defer to the .searchsorted method if it is present

Thanks, didn't realize that. I went down the __array_function__ rabbit hole for a good while. Too bad the numpy docs don't have a little note in those docs to the effect of "if you are dealing with an ndarray subclass you might not need this".